MdePkg: add an instance of UnitTestHostBaseLib for AArch64
Add an instance of UnitTestHostBaseLib for AArch64 so that
the Host Based Unit test support can be enabled for AArch64.
