+++ MdePkg/Library/BaseLib/UnitTestHostBaseLib.inf
@@ -23,7 +23,7 @@
   LIBRARY_CLASS                  = UnitTestHostBaseLib|HOST_APPLICATION
 
 #
-#  VALID_ARCHITECTURES           = IA32 X64
+#  VALID_ARCHITECTURES           = IA32 X64 AARCH64
 #
 
 [Sources]
@@ -182,9 +182,10 @@
   Math64.c
 
 [Sources.AARCH64]
+  AArch64UnitTestHost.c
   AArch64/InternalSwitchStack.c
-  AArch64/Unaligned.c
   Math64.c
+  Unaligned.c
 
   AArch64/MemoryFence.S             | GCC
   AArch64/SwitchStack.S             | GCC
